Renamed processor from filter to pass
