+++ plugins/processors/all/all.go
@@ -21,7 +21,7 @@ import (
 	_ "github.com/influxdata/telegraf/plugins/processors/s2geo"
 	_ "github.com/influxdata/telegraf/plugins/processors/starlark"
 	_ "github.com/influxdata/telegraf/plugins/processors/strings"
-	_ "github.com/influxdata/telegraf/plugins/processors/t128_filter"
+	_ "github.com/influxdata/telegraf/plugins/processors/t128_pass"
 	_ "github.com/influxdata/telegraf/plugins/processors/t128_transform"
 	_ "github.com/influxdata/telegraf/plugins/processors/tag_limit"
 	_ "github.com/influxdata/telegraf/plugins/processors/template"
